+++ controllers/contacts/changeStatusContact.js
@@ -1,0 +1,14 @@
+const HttpError = require("../../helpers");
+const { Contact } = require("../../models");
+const { ctrlWrapper } = require("../../utils");
+
+const changeStatusContact = async (req, res) => {
+  const { id } = req.params;
+  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  if (!result) {
+    throw HttpError(404, `Contact with ${id} not found `);
+  }
+  res.json(result);
+};
+
+module.exports = { changeStatusContact: ctrlWrapper(changeStatusContact) };
