+++ models/index.js
@@ -1,0 +1,6 @@
+const { Contact, schemas } = require("./contacts");
+
+module.exports = {
+  Contact,
+  schemas,
+};
